LayoutTests/fast/dom/MicroData/properties-collection-namedgetter-with-invalid-name-expected.txt

Test to ensure that HTMLPropertiesCollection[name] should return undefined if item doesn't have a property named 'name'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testDiv.properties['foo'] is undefined
PASS !testDiv.properties['foo'] is true
PASS testDiv.properties['foo'] == '[object PropertyNodeList]' is true
PASS testDiv.properties['bar'] is undefined
PASS !testDiv.properties['bar'] is true
PASS successfullyParsed is true

TEST COMPLETE
